requirements.txt
annotated-types==0.7.0
fastapi==0.115.12
openpyxl==3.1.5
oracledb==3.1.1
pandas==2.2.3
polars==0.20.31
pycparser==2.22
pydantic==2.11.5
pydantic-core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
typing-extensions==4.13.2
typing-inspection==0.4.1
tzdata==2025.2
uvicorn==0.34.2
polars~=0.20
SQLAlchemy~=2.0
PyMySQL~=1.1
pandas~=2.2
pyarrow

